Refactors the OpenSearch sink to split out the ingestion work from other work such as index management. I plan to use this work to help support pull-based ingestion within this sink. This commit is just for refactoring since it is already significant.
