+++ data-prepper-plugins/opensearch/src/main/java/org/opensearch/dataprepper/plugins/sink/opensearch/OpenSearchSink.java
@@ -9,101 +9,43 @@
 
 package org.opensearch.dataprepper.plugins.sink.opensearch;
 
-import com.google.common.annotations.VisibleForTesting;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import io.micrometer.core.instrument.Counter;
-import io.micrometer.core.instrument.DistributionSummary;
-import io.micrometer.core.instrument.Timer;
-import org.apache.commons.lang3.StringUtils;
 import org.opensearch.client.RestHighLevelClient;
 import org.opensearch.client.opensearch.OpenSearchClient;
-import org.opensearch.client.opensearch._types.VersionType;
-import org.opensearch.client.opensearch.core.BulkRequest;
-import org.opensearch.client.opensearch.core.bulk.BulkOperation;
 import org.opensearch.client.transport.TransportOptions;
-import org.opensearch.common.unit.ByteSizeUnit;
 import org.opensearch.dataprepper.aws.api.AwsCredentialsSupplier;
-import org.opensearch.dataprepper.common.concurrent.BackgroundThreadFactory;
-import org.opensearch.dataprepper.expression.ExpressionEvaluationException;
 import org.opensearch.dataprepper.expression.ExpressionEvaluator;
-import org.opensearch.dataprepper.metrics.MetricNames;
 import org.opensearch.dataprepper.model.annotations.DataPrepperPlugin;
 import org.opensearch.dataprepper.model.annotations.DataPrepperPluginConstructor;
 import org.opensearch.dataprepper.model.configuration.PipelineDescription;
 import org.opensearch.dataprepper.model.configuration.PluginSetting;
 import org.opensearch.dataprepper.model.event.Event;
-import org.opensearch.dataprepper.model.event.EventHandle;
-import org.opensearch.dataprepper.model.event.InternalEventHandle;
-import org.opensearch.dataprepper.model.event.exceptions.EventKeyNotFoundException;
-import org.opensearch.dataprepper.model.failures.DlqObject;
-import org.opensearch.dataprepper.model.opensearch.OpenSearchBulkActions;
 import org.opensearch.dataprepper.model.plugin.InvalidPluginConfigurationException;
 import org.opensearch.dataprepper.model.plugin.PluginConfigObservable;
 import org.opensearch.dataprepper.model.record.Record;
 import org.opensearch.dataprepper.model.sink.AbstractSink;
 import org.opensearch.dataprepper.model.sink.Sink;
 import org.opensearch.dataprepper.model.sink.SinkContext;
-import org.opensearch.dataprepper.model.sink.SinkForwardRecordsContext;
 import org.opensearch.dataprepper.plugins.common.opensearch.ServerlessNetworkPolicyUpdater;
 import org.opensearch.dataprepper.plugins.common.opensearch.ServerlessNetworkPolicyUpdaterFactory;
 import org.opensearch.dataprepper.plugins.common.opensearch.ServerlessOptionsFactory;
-import org.opensearch.dataprepper.plugins.dlq.DlqProvider;
-import org.opensearch.dataprepper.plugins.dlq.DlqWriter;
-import org.opensearch.dataprepper.plugins.dlq.s3.S3DlqProvider;
-import org.opensearch.dataprepper.plugins.sink.opensearch.bulk.AccumulatingBulkRequest;
-import org.opensearch.dataprepper.plugins.sink.opensearch.bulk.BulkApiWrapper;
-import org.opensearch.dataprepper.plugins.sink.opensearch.bulk.BulkApiWrapperFactory;
-import org.opensearch.dataprepper.plugins.sink.opensearch.bulk.BulkOperationWriter;
-import org.opensearch.dataprepper.plugins.sink.opensearch.bulk.JavaClientAccumulatingCompressedBulkRequest;
-import org.opensearch.dataprepper.plugins.sink.opensearch.bulk.JavaClientAccumulatingUncompressedBulkRequest;
-import org.opensearch.dataprepper.plugins.sink.opensearch.bulk.SerializedJson;
-import org.opensearch.dataprepper.plugins.sink.opensearch.configuration.ActionConfiguration;
-import org.opensearch.dataprepper.plugins.sink.opensearch.configuration.DlqConfiguration;
 import org.opensearch.dataprepper.plugins.sink.opensearch.configuration.OpenSearchSinkConfig;
-import org.opensearch.dataprepper.plugins.sink.opensearch.dlq.FailedBulkOperation;
-import org.opensearch.dataprepper.plugins.sink.opensearch.dlq.FailedBulkOperationConverter;
-import org.opensearch.dataprepper.plugins.sink.opensearch.dlq.FailedDlqData;
 import org.opensearch.dataprepper.plugins.sink.opensearch.index.ClusterSettingsParser;
-import org.opensearch.dataprepper.plugins.sink.opensearch.index.DataStreamDetector;
-import org.opensearch.dataprepper.plugins.sink.opensearch.index.DataStreamIndex;
-import org.opensearch.dataprepper.plugins.sink.opensearch.index.IndexCache;
-import org.opensearch.dataprepper.plugins.sink.opensearch.index.DocumentBuilder;
-import org.opensearch.dataprepper.plugins.sink.opensearch.index.ExistingDocumentQueryManager;
+import org.opensearch.dataprepper.plugins.sink.opensearch.index.CustomDocumentBuilderFactory;
 import org.opensearch.dataprepper.plugins.sink.opensearch.index.IndexManager;
 import org.opensearch.dataprepper.plugins.sink.opensearch.index.IndexManagerFactory;
 import org.opensearch.dataprepper.plugins.sink.opensearch.index.IndexTemplateAPIWrapper;
 import org.opensearch.dataprepper.plugins.sink.opensearch.index.IndexTemplateAPIWrapperFactory;
 import org.opensearch.dataprepper.plugins.sink.opensearch.index.IndexType;
-import org.opensearch.dataprepper.plugins.sink.opensearch.index.CustomDocumentBuilder;
-import org.opensearch.dataprepper.plugins.sink.opensearch.index.CustomDocumentBuilderFactory;
 import org.opensearch.dataprepper.plugins.sink.opensearch.index.TemplateStrategy;
 import org.opensearch.dataprepper.plugins.source.opensearch.configuration.ServerlessOptions;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.BufferedWriter;
 import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Paths;
-import java.nio.file.StandardOpenOption;
-import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Objects;
 import java.util.Optional;
-import java.util.Set;
-import java.util.StringJoiner;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-import java.util.concurrent.locks.ReentrantLock;
 import java.util.function.Function;
-import java.util.function.Supplier;
-import java.util.stream.Collectors;
-
-import static org.opensearch.dataprepper.logging.DataPrepperMarkers.NOISY;
 
 @DataPrepperPlugin(name = "opensearch", pluginType = Sink.class, pluginConfigurationType = OpenSearchSinkConfig.class)
 public class OpenSearchSink extends AbstractSink<Record<Event>> {
@@ -113,71 +55,22 @@ public class OpenSearchSink extends AbstractSink<Record<Event>> {
   public static final String BULKREQUEST_SIZE_BYTES = "bulkRequestSizeBytes";
   public static final String DYNAMIC_INDEX_DROPPED_EVENTS = "dynamicIndexDroppedEvents";
   public static final String INVALID_VERSION_EXPRESSION_DROPPED_EVENTS = "dynamicDocumentVersionDroppedEvents";
-  private static final String PLUGIN_NAME = "opensearch";
 
   private static final Logger LOG = LoggerFactory.getLogger(OpenSearchSink.class);
   private static final int INITIALIZE_RETRY_WAIT_TIME_MS = 5000;
-  private final AwsCredentialsSupplier awsCredentialsSupplier;
 
-  private DlqWriter dlqWriter;
-  private BufferedWriter dlqFileWriter;
+  private final AwsCredentialsSupplier awsCredentialsSupplier;
   private final OpenSearchSinkConfiguration openSearchSinkConfig;
   private final IndexManagerFactory indexManagerFactory;
-  private RestHighLevelClient restHighLevelClient;
-  private IndexManager indexManager;
-  private Supplier<AccumulatingBulkRequest> bulkRequestSupplier;
-  private BulkRetryStrategy bulkRetryStrategy;
-  private BulkApiWrapper bulkApiWrapper;
-  private final long bulkSize;
-  private final long flushTimeout;
   private final IndexType indexType;
-  private final String documentIdField;
-  private final String documentId;
-  private final String routingField;
-  private final String routing;
-  private final String pipeline;
-  private final String action;
-  private final List<ActionConfiguration> actions;
-  private final ScriptManager scriptManager;
-  private final BulkOperationFactory bulkOperationFactory;
-  private final String documentRootKey;
-  private String configuredIndexAlias;
-  private final ReentrantLock lock;
-  private final VersionType versionType;
-  private final String versionExpression;
-
-  private final Timer bulkRequestTimer;
-  private final Counter bulkRequestErrorsCounter;
-  private final Counter invalidActionErrorsCounter;
-  private final Counter dynamicIndexDroppedEvents;
-  private final DistributionSummary bulkRequestSizeBytesSummary;
-  private final Counter dynamicDocumentVersionDroppedEvents;
-  private OpenSearchClient openSearchClient;
-  private OpenSearchClientRefresher openSearchClientRefresher;
-  private ObjectMapper objectMapper;
-  private volatile boolean initialized;
-  private final SinkContext sinkContext;
-  private final ExpressionEvaluator expressionEvaluator;
-  private boolean useEventInBulkOperation;
-
-  private FailedBulkOperationConverter failedBulkOperationConverter;
-  private DataStreamDetector dataStreamDetector;
-  private DataStreamIndex dataStreamIndex;
-  IndexCache indexCache;
-
-  private DlqProvider dlqProvider;
-  private final ConcurrentHashMap<Long, AccumulatingBulkRequest<BulkOperationWrapper, BulkRequest>> bulkRequestMap;
-  private final ConcurrentHashMap<Long, Long> lastFlushTimeMap;
   private final PluginConfigObservable pluginConfigObservable;
+  private final Ingester ingester;
 
-  private ExistingDocumentQueryManager existingDocumentQueryManager;
-
-  private final CustomDocumentBuilder customDocumentBuilder;
-
-  private final ExecutorService queryExecutorService;
-
-  private final int processWorkerThreads;
-  private final SinkForwardRecordsContext sinkForwardRecordsContext;
+  private final RestHighLevelClient restHighLevelClient;
+  private final OpenSearchClient openSearchClient;
+  private final OpenSearchClientRefresher openSearchClientRefresher;
+  private IndexManager indexManager;
+  private volatile boolean initialized;
 
   @DataPrepperPluginConstructor
   public OpenSearchSink(final PluginSetting pluginSetting,
@@ -188,60 +81,48 @@ public class OpenSearchSink extends AbstractSink<Record<Event>> {
                         final PluginConfigObservable pluginConfigObservable,
                         final OpenSearchSinkConfig openSearchSinkConfiguration) {
     super(pluginSetting, Integer.MAX_VALUE, INITIALIZE_RETRY_WAIT_TIME_MS);
-    this.processWorkerThreads = pipelineDescription.getNumberOfProcessWorkers();
     this.awsCredentialsSupplier = awsCredentialsSupplier;
-    this.sinkContext = sinkContext != null ? sinkContext : new SinkContext(null, Collections.emptyList(), Collections.emptyList(), Collections.emptyList());
-    sinkForwardRecordsContext = new SinkForwardRecordsContext(sinkContext);
-    this.expressionEvaluator = expressionEvaluator;
-    this.pipeline = pipelineDescription.getPipelineName();
-    bulkRequestTimer = pluginMetrics.timer(BULKREQUEST_LATENCY);
-    bulkRequestErrorsCounter = pluginMetrics.counter(BULKREQUEST_ERRORS);
-    invalidActionErrorsCounter = pluginMetrics.counter(INVALID_ACTION_ERRORS);
-    dynamicIndexDroppedEvents = pluginMetrics.counter(DYNAMIC_INDEX_DROPPED_EVENTS);
-    bulkRequestSizeBytesSummary = pluginMetrics.summary(BULKREQUEST_SIZE_BYTES);
-    dynamicDocumentVersionDroppedEvents = pluginMetrics.counter(INVALID_VERSION_EXPRESSION_DROPPED_EVENTS);
+    this.pluginConfigObservable = pluginConfigObservable;
+
+    final SinkContext resolvedSinkContext = sinkContext != null ? sinkContext :
+            new SinkContext(null, Collections.emptyList(), Collections.emptyList(), Collections.emptyList());
 
     this.openSearchSinkConfig = OpenSearchSinkConfiguration.readOSConfig(openSearchSinkConfiguration, expressionEvaluator);
-    this.bulkSize = ByteSizeUnit.MB.toBytes(openSearchSinkConfig.getIndexConfiguration().getBulkSize());
-    this.flushTimeout = openSearchSinkConfig.getIndexConfiguration().getFlushTimeout();
     this.indexType = openSearchSinkConfig.getIndexConfiguration().getIndexType();
-    this.documentIdField = openSearchSinkConfig.getIndexConfiguration().getDocumentIdField();
-    this.documentId = openSearchSinkConfig.getIndexConfiguration().getDocumentId();
-    this.routingField = openSearchSinkConfig.getIndexConfiguration().getRoutingField();
-    this.routing = openSearchSinkConfig.getIndexConfiguration().getRouting();
-    this.action = openSearchSinkConfig.getIndexConfiguration().getAction();
-    this.actions = openSearchSinkConfig.getIndexConfiguration().getActions();
-    this.scriptManager = new ScriptManager(openSearchSinkConfig.getIndexConfiguration().getScriptConfiguration(), expressionEvaluator);
-    this.documentRootKey = openSearchSinkConfig.getIndexConfiguration().getDocumentRootKey();
-    this.versionType = openSearchSinkConfig.getIndexConfiguration().getVersionType();
-    this.versionExpression = openSearchSinkConfig.getIndexConfiguration().getVersionExpression();
     this.indexManagerFactory = new IndexManagerFactory(new ClusterSettingsParser());
-    this.failedBulkOperationConverter = new FailedBulkOperationConverter(pipeline, PLUGIN_NAME);
     this.initialized = false;
-    this.lock = new ReentrantLock(true);
-    this.bulkRequestMap = new ConcurrentHashMap<>();
-    this.lastFlushTimeMap = new ConcurrentHashMap<>();
-    this.pluginConfigObservable = pluginConfigObservable;
-    this.objectMapper = new ObjectMapper();
-    this.bulkOperationFactory = new BulkOperationFactory(versionType, scriptManager, objectMapper, isUsingDocumentFilters());
-    this.customDocumentBuilder = new CustomDocumentBuilderFactory().create(this.indexType);
-    this.queryExecutorService = openSearchSinkConfig.getIndexConfiguration().getQueryTerm() != null ?
-            Executors.newSingleThreadExecutor(BackgroundThreadFactory.defaultExecutorThreadFactory("existing-document-query-manager")) : null;
 
-    final Optional<DlqConfiguration> dlqConfig = openSearchSinkConfig.getRetryConfiguration().getDlq();
+    final ConnectionConfiguration connectionConfiguration = openSearchSinkConfig.getConnectionConfiguration();
+    restHighLevelClient = connectionConfiguration.createClient(awsCredentialsSupplier);
+    openSearchClient = connectionConfiguration.createOpenSearchClient(restHighLevelClient, awsCredentialsSupplier);
+    final Function<ConnectionConfiguration, OpenSearchClient> clientFunction =
+            (connConfig) -> {
+      final RestHighLevelClient client = connConfig.createClient(awsCredentialsSupplier);
+      return connConfig.createOpenSearchClient(client, awsCredentialsSupplier).withTransportOptions(
+              TransportOptions.builder()
+                      .setParameter("filter_path", "errors,took,items.*.error,items.*.status,items.*._index,items.*._id")
+                      .build());
+    };
+    openSearchClientRefresher = new OpenSearchClientRefresher(
+            pluginMetrics, connectionConfiguration, clientFunction);
 
-    if (dlqConfig.isPresent()) {
-      dlqProvider = new S3DlqProvider(dlqConfig.get().getS3DlqWriterConfig());
-    }
+    final String pipeline = pipelineDescription.getPipelineName();
+    final EventActionResolver eventActionResolver = new EventActionResolver(
+            openSearchSinkConfig.getIndexConfiguration().getAction(),
+            openSearchSinkConfig.getIndexConfiguration().getActions(),
+            expressionEvaluator);
+
+    this.ingester = new BulkIngester(openSearchSinkConfig, expressionEvaluator, resolvedSinkContext,
+            pluginMetrics, pipeline, eventActionResolver,
+            openSearchClient, () -> openSearchClientRefresher.get(),
+            this::getIndexManager, this::getFailurePipeline,
+            new CustomDocumentBuilderFactory().create(this.indexType));
   }
 
   @Override
   public void doInitialize() {
     try {
         doInitializeInternal();
-        // getFailurePipeline() does not return valid value in the constructor. Earliest it can be used
-        // is in doInitialize()
-        useEventInBulkOperation = (getFailurePipeline() != null || sinkContext.getForwardToPipelines().size() > 0);
     } catch (IOException e) {
         LOG.warn("Failed to initialize OpenSearch sink, retrying: {} ", e.getMessage());
         this.shutdown();
@@ -261,87 +142,30 @@ public class OpenSearchSink extends AbstractSink<Record<Event>> {
 
   private void doInitializeInternal() throws IOException {
     LOG.info("Initializing OpenSearch sink");
-    final ConnectionConfiguration connectionConfiguration = openSearchSinkConfig.getConnectionConfiguration();
-    restHighLevelClient = connectionConfiguration.createClient(awsCredentialsSupplier);
-    openSearchClient = connectionConfiguration.createOpenSearchClient(restHighLevelClient, awsCredentialsSupplier);
-    final Function<ConnectionConfiguration, OpenSearchClient> clientFunction =
-            (connectionConfiguration1) -> {
-      final RestHighLevelClient restHighLevelClient1 = connectionConfiguration1.createClient(awsCredentialsSupplier);
-      return connectionConfiguration1.createOpenSearchClient(restHighLevelClient1, awsCredentialsSupplier).withTransportOptions(
-              TransportOptions.builder()
-                      .setParameter("filter_path", "errors,took,items.*.error,items.*.status,items.*._index,items.*._id")
-                      .build());
-    };
-    openSearchClientRefresher = new OpenSearchClientRefresher(
-            pluginMetrics, connectionConfiguration, clientFunction);
 
     pluginConfigObservable.addPluginConfigObserver(
             newOpenSearchSinkConfig -> openSearchClientRefresher.update((OpenSearchSinkConfig) newOpenSearchSinkConfig));
-    configuredIndexAlias = openSearchSinkConfig.getIndexConfiguration().getIndexAlias();
+
+    final String configuredIndexAlias = openSearchSinkConfig.getIndexConfiguration().getIndexAlias();
     final IndexTemplateAPIWrapper indexTemplateAPIWrapper = IndexTemplateAPIWrapperFactory.getWrapper(
             openSearchSinkConfig.getIndexConfiguration(), openSearchClient);
     final TemplateStrategy templateStrategy = openSearchSinkConfig.getIndexConfiguration().getTemplateType()
             .createTemplateStrategy(indexTemplateAPIWrapper);
     indexManager = indexManagerFactory.getIndexManager(indexType, openSearchClient, restHighLevelClient,
             openSearchSinkConfig, templateStrategy, configuredIndexAlias);
-    final String dlqFile = openSearchSinkConfig.getRetryConfiguration().getDlqFile();
-     if (dlqFile != null) {
-      dlqFileWriter = Files.newBufferedWriter(Paths.get(dlqFile), StandardOpenOption.CREATE, StandardOpenOption.APPEND);
-    } else if (dlqProvider != null) {
-      Optional<DlqWriter> potentialDlq = dlqProvider.getDlqWriter(new StringJoiner(MetricNames.DELIMITER)
-          .add(pipeline)
-          .add(PLUGIN_NAME).toString());
-      dlqWriter = potentialDlq.isPresent() ? potentialDlq.get() : null;
-    }
 
-    // Attempt to update the serverless network policy if required argument are given.
     maybeUpdateServerlessNetworkPolicy();
 
     indexManager.setupIndex();
 
-    final Boolean requireAlias = indexManager.isIndexAlias(configuredIndexAlias);
-    final boolean isEstimateBulkSizeUsingCompression = openSearchSinkConfig.getIndexConfiguration().isEstimateBulkSizeUsingCompression();
-    final boolean isRequestCompressionEnabled = openSearchSinkConfig.getConnectionConfiguration().isRequestCompressionEnabled();
-    if (isEstimateBulkSizeUsingCompression && isRequestCompressionEnabled) {
-      final int maxLocalCompressionsForEstimation = openSearchSinkConfig.getIndexConfiguration().getMaxLocalCompressionsForEstimation();
-      bulkRequestSupplier = () -> new JavaClientAccumulatingCompressedBulkRequest(new BulkRequest.Builder().requireAlias(requireAlias), bulkSize, maxLocalCompressionsForEstimation);
-    } else if (isEstimateBulkSizeUsingCompression) {
-      LOG.warn("Estimate bulk request size using compression was enabled but request compression is disabled. " +
-              "Estimating bulk request size without compression.");
-      bulkRequestSupplier = () -> new JavaClientAccumulatingUncompressedBulkRequest(new BulkRequest.Builder().requireAlias(requireAlias));
-    } else {
-      bulkRequestSupplier = () -> new JavaClientAccumulatingUncompressedBulkRequest(new BulkRequest.Builder().requireAlias(requireAlias));
-    }
-
-    final int maxRetries = openSearchSinkConfig.getRetryConfiguration().getMaxRetries();
-    bulkApiWrapper = BulkApiWrapperFactory.getWrapper(openSearchSinkConfig.getIndexConfiguration(),
-            () -> openSearchClientRefresher.get());
-    bulkRetryStrategy = new BulkRetryStrategy(bulkRequest -> bulkApiWrapper.bulk(bulkRequest.getRequest()),
-            this::logFailureForBulkRequests,
-            this::successfulOperationsHandler,
-            pluginMetrics,
-            maxRetries,
-            bulkRequestSupplier,
-            pipeline,
-            PLUGIN_NAME,
-            openSearchSinkConfig.getIndexConfiguration().getQueryOnBulkFailures() ? existingDocumentQueryManager : null,
-            isExternalVersionType(openSearchSinkConfig.getIndexConfiguration().getVersionType()));
-
-    if (queryExecutorService != null) {
-      existingDocumentQueryManager = new ExistingDocumentQueryManager(openSearchSinkConfig.getIndexConfiguration(), pluginMetrics, openSearchClient);
-      queryExecutorService.submit(existingDocumentQueryManager);
-    }
-
-    this.indexCache = new IndexCache();
-    this.dataStreamDetector = new DataStreamDetector(openSearchClient, indexCache);
-    this.dataStreamIndex = new DataStreamIndex(dataStreamDetector, openSearchSinkConfig.getIndexConfiguration());
+    ingester.initialize();
 
     this.initialized = true;
     LOG.info("Initialized OpenSearch sink");
   }
 
-  double getInvalidActionErrorsCount() {
-    return invalidActionErrorsCounter.count();
+  IndexManager getIndexManager() {
+    return indexManager;
   }
 
   @Override
@@ -351,297 +175,13 @@ public class OpenSearchSink extends AbstractSink<Record<Event>> {
 
   @Override
   public void doOutput(final Collection<Record<Event>> records) {
-    final long threadId = Thread.currentThread().getId();
-    if (!bulkRequestMap.containsKey(threadId)) {
-      bulkRequestMap.put(threadId, bulkRequestSupplier.get());
-    }
-    if (!lastFlushTimeMap.containsKey(threadId)) {
-      lastFlushTimeMap.put(threadId, System.currentTimeMillis());
-    }
-
-    AccumulatingBulkRequest<BulkOperationWrapper, BulkRequest> bulkRequest = bulkRequestMap.get(threadId);
-    long lastFlushTime = lastFlushTimeMap.get(threadId);
-
-    Set<BulkOperationWrapper> documentsReadyForIndexing = new HashSet<>();
-    if (openSearchSinkConfig.getIndexConfiguration().getQueryTerm() != null) {
-      documentsReadyForIndexing = existingDocumentQueryManager.getAndClearBulkOperationsReadyToIndex();
-    }
-
-
-    if (!documentsReadyForIndexing.isEmpty()) {
-      LOG.info("Found {} documents ready for indexing from query manager", documentsReadyForIndexing.size());
-    }
-
-    for (final BulkOperationWrapper bulkOperationWrapper : documentsReadyForIndexing) {
-      bulkRequest = flushBatch(bulkRequest, bulkOperationWrapper, lastFlushTime);
-      bulkRequest.addOperation(bulkOperationWrapper);
-    }
-
-
-    for (final Record<Event> record : records) {
-      final Event event = record.getData();
-      String indexName = configuredIndexAlias;
-      try {
-          indexName = indexManager.getIndexName(event.formatString(indexName, expressionEvaluator));
-      } catch (final Exception e) {
-          LOG.error(NOISY, "There was an exception when constructing the index name. Check the dlq if configured to see details about the affected Event: {}", e.getMessage());
-          dynamicIndexDroppedEvents.increment();
-          logFailureForDlqObjects(List.of(createDlqObjectFromEvent(event, indexName, e.getMessage())), e);
-          continue;
-      }
-      
-      dataStreamIndex.ensureTimestamp(event, indexName);
-
-      if (customDocumentBuilder != null) {
-        try {
-          final List<String> tsdbDocs = customDocumentBuilder.buildDocuments(event);
-          final String tsdbAction = resolveEventAction(event);
-          final EventHandle eventHandle = event.getEventHandle();
-          if (tsdbDocs.size() > 1 && eventHandle instanceof InternalEventHandle) {
-            for (int i = 0; i < tsdbDocs.size() - 1; i++) {
-              ((InternalEventHandle) eventHandle).acquireReference();
-            }
-          }
-          for (final String tsdbDoc : tsdbDocs) {
-            final SerializedJson doc = SerializedJson.fromStringAndOptionals(tsdbDoc, null, null, null);
-            final BulkOperation op = bulkOperationFactory.create(tsdbAction, doc, null, indexName, null);
-            final BulkOperationWrapper wrapper = new BulkOperationWrapper(op, eventHandle, null, null);
-            bulkRequest = flushBatch(bulkRequest, wrapper, lastFlushTime);
-            bulkRequest.addOperation(wrapper);
-          }
-        } catch (final Exception e) {
-          LOG.error("Failed to build TSDB documents for event: {}", e.getMessage(), e);
-          dynamicIndexDroppedEvents.increment();
-          logFailureForDlqObjects(List.of(createDlqObjectFromEvent(event, indexName, e.getMessage())), e);
-        }
-        continue;
-      }
-
-      final SerializedJson document = getDocument(event);
-
-      Long version = null;
-      String versionExpressionEvaluationResult = null;
-      if (versionExpression != null) {
-        try {
-          versionExpressionEvaluationResult = event.formatString(versionExpression, expressionEvaluator);
-          version = Long.valueOf(event.formatString(versionExpression, expressionEvaluator));
-        } catch (final NumberFormatException e) {
-          // Skip this event from the bulk request after sending to DLQ to avoid
-          // including events with invalid versions in the OpenSearch bulk request.
-          // See https://github.com/opensearch-project/data-prepper/issues/6601
-          final String errorMessage = String.format(
-                  "Unable to convert the result of evaluating document_version '%s' to Long for an Event. The evaluation result '%s' must be a valid Long type", versionExpression, versionExpressionEvaluationResult
-          );
-          LOG.error(errorMessage);
-          logFailureForDlqObjects(List.of(createDlqObjectFromEvent(event, indexName, errorMessage)), e);
-          dynamicDocumentVersionDroppedEvents.increment();
-          continue;
-        } catch (final RuntimeException e) {
-          // Skip this event from the bulk request after sending to DLQ to avoid
-          // including events with invalid versions in the OpenSearch bulk request.
-          // See https://github.com/opensearch-project/data-prepper/issues/6601
-          final String errorMessage = String.format(
-                  "There was an exception when evaluating the document_version '%s': %s", versionExpression, e.getMessage());
-          LOG.error(errorMessage + " Check the dlq if configured to see more details about the affected Event");
-          logFailureForDlqObjects(List.of(createDlqObjectFromEvent(event, indexName, errorMessage)), e);
-          dynamicDocumentVersionDroppedEvents.increment();
-          continue;
-        }
-      }
-
-      String eventAction = resolveEventAction(event);
-      
-      if (dataStreamDetector.isDataStream(indexName)) {
-        eventAction = dataStreamIndex.determineAction(eventAction, indexName);
-      }
-      if (OpenSearchBulkActions.fromOptionValue(eventAction) == null) {
-        LOG.error("Unknown action {}, skipping the event", eventAction);
-        invalidActionErrorsCounter.increment();
-        continue;
-      }
-
-      SerializedJson serializedJsonNode = null;
-      if (StringUtils.equals(eventAction, OpenSearchBulkActions.UPDATE.toString()) ||
-          StringUtils.equals(eventAction, OpenSearchBulkActions.UPSERT.toString()) ||
-          StringUtils.equals(eventAction, OpenSearchBulkActions.DELETE.toString())) {
-            serializedJsonNode = SerializedJson.fromJsonNode(event.getJsonNode(), document);
-      }
-      BulkOperation bulkOperation;
-
-      try {
-        bulkOperation = bulkOperationFactory.create(eventAction, document, version, indexName, event.getJsonNode());
-      } catch (final Exception e) {
-        LOG.error("An exception occurred while constructing the bulk operation for a document: ", e);
-        logFailureForDlqObjects(List.of(createDlqObjectFromEvent(event, indexName, e.getMessage())), e);
-        continue;
-      }
-
-      final String queryTermKey = openSearchSinkConfig.getIndexConfiguration().getQueryTerm();
-      final String termValue = queryTermKey != null ?
-              event.get(queryTermKey, String.class) : null;
-      BulkOperationWrapper bulkOperationWrapper = (useEventInBulkOperation) ?
-              new BulkOperationWrapper(bulkOperation, event, serializedJsonNode, termValue) :
-              new BulkOperationWrapper(bulkOperation, event.getEventHandle(), serializedJsonNode, termValue);
-
-      if (openSearchSinkConfig.getIndexConfiguration().getQueryWhen() != null &&
-              expressionEvaluator.evaluateConditional(openSearchSinkConfig.getIndexConfiguration().getQueryWhen(), event)) {
-        existingDocumentQueryManager.addBulkOperation(bulkOperationWrapper);
-        continue;
-      }
-
-      bulkRequest = flushBatch(bulkRequest, bulkOperationWrapper, lastFlushTime);
-      bulkRequest.addOperation(bulkOperationWrapper);
-    }
-
-    // Flush the remaining requests if flush timeout expired
-    if (System.currentTimeMillis() - lastFlushTime > flushTimeout && bulkRequest.getOperationsCount() > 0) {
-      flushBatch(bulkRequest);
-      lastFlushTime = System.currentTimeMillis();
-      bulkRequest = bulkRequestSupplier.get();
-    }
-
-    bulkRequestMap.put(threadId, bulkRequest);
-    lastFlushTimeMap.put(threadId, lastFlushTime);
+    ingester.output(records);
   }
 
-  SerializedJson getDocument(final Event event) {
-    String docId = null;
-
-    if (Objects.nonNull(documentIdField)) {
-      docId = event.get(documentIdField, String.class);
-    } else if (Objects.nonNull(documentId)) {
-      try {
-        docId = event.formatString(documentId, expressionEvaluator);
-      } catch (final ExpressionEvaluationException | EventKeyNotFoundException e) {
-        LOG.error("Unable to construct document_id with format {}, the document_id will be generated by OpenSearch", documentId, e);
-      }
-    }
-
-    String routingValue = null;
-    if (routingField != null) {
-      routingValue = event.get(routingField, String.class);
-    } else if (routing != null) {
-      try {
-        routingValue = event.formatString(routing, expressionEvaluator);
-      } catch (final ExpressionEvaluationException | EventKeyNotFoundException e) {
-        LOG.error("Unable to construct routing with format {}, the routing will be generated by OpenSearch", routing, e);
-      }
-    }
-
-    final String document = DocumentBuilder.build(event, documentRootKey, sinkContext.getTagsTargetKey(), sinkContext.getIncludeKeys(), sinkContext.getExcludeKeys());
-
-    return SerializedJson.builder()
-            .withJsonString(document)
-            .withDocumentId(docId)
-            .withRoutingField(routingValue)
-            .withResolvedScriptParameters(scriptManager.resolveParams(event))
-            .build();
-  }
-
-  private void flushBatch(AccumulatingBulkRequest accumulatingBulkRequest) {
-    bulkRequestTimer.record(() -> {
-      try {
-        LOG.debug("Sending data to OpenSearch");
-        bulkRetryStrategy.execute(accumulatingBulkRequest);
-        bulkRequestSizeBytesSummary.record(accumulatingBulkRequest.getEstimatedSizeInBytes());
-      } catch (final InterruptedException e) {
-        LOG.error("Unexpected Interrupt:", e);
-        bulkRequestErrorsCounter.increment();
-        Thread.currentThread().interrupt();
-      }
-    });
-  }
-
-  @VisibleForTesting
-  void successfulOperationsHandler(final List<BulkOperationWrapper> successfulOperations) {
-    if (successfulOperations.size() == 0) {
-        return;
-    }
-    if (sinkContext.getForwardToPipelines().size() == 0) {
-      for (final BulkOperationWrapper bulkOperation: successfulOperations) {
-        if (bulkOperation.getEvent() != null) {
-            bulkOperation.getEvent().getEventHandle().release(true);
-        } else {
-            bulkOperation.releaseEventHandle(true);
-        }
-      }
-      return;
-    }
-    for (final BulkOperationWrapper bulkOperation: successfulOperations) {
-      sinkForwardRecordsContext.addRecord(new Record<>(bulkOperation.getEvent()));
-    }
-    sinkContext.forwardRecords(sinkForwardRecordsContext, null, null);
-  }
-
-  private void logFailureForBulkRequests(final List<FailedBulkOperation> failedBulkOperations, final Throwable failure) {
-
-    final List<DlqObject> dlqObjects = failedBulkOperations.stream()
-        .map(failedBulkOperationConverter::convertToDlqObject)
-        .collect(Collectors.toList());
-
-    logFailureForDlqObjects(dlqObjects, failure);
-  }
-
-  private void logFailureForDlqObjects(final List<DlqObject> dlqObjects, final Throwable failure) {
-    if (getFailurePipeline() != null) {
-      List<Record<Event>> records = new ArrayList<>();
-      for (DlqObject dlqObject : dlqObjects) {
-        Event event = dlqObject.getEvent();
-
-        if (event != null) {
-            event.updateFailureMetadata()
-                .withPluginId(dlqObject.getPluginId())
-                .withPluginName(dlqObject.getPluginName())
-                .withPipelineName(dlqObject.getPipelineName())
-                .withErrorMessage(((FailedDlqData) dlqObject.getFailedData()).getMessage())
-                .with("status", ((FailedDlqData) dlqObject.getFailedData()).getStatus())
-                .with("index", ((FailedDlqData) dlqObject.getFailedData()).getIndex())
-                .with("indexId", ((FailedDlqData) dlqObject.getFailedData()).getIndexId());
-          records.add(new Record<>(event));
-        }
-      }
-      getFailurePipeline().sendEvents(records);
-      return;
-    }
-    if (dlqFileWriter != null) {
-      dlqObjects.forEach(dlqObject -> {
-        final FailedDlqData failedDlqData = (FailedDlqData) dlqObject.getFailedData();
-        final String message = failure == null ? failedDlqData.getMessage() : failure.getMessage();
-        try {
-          dlqFileWriter.write(String.format("{\"Document\": [%s], \"failure\": %s}\n",
-                  BulkOperationWriter.dlqObjectToString(dlqObject), message));
-          dlqObject.releaseEventHandle(true);
-        } catch (final IOException e) {
-        LOG.error(NOISY, "Failed to write a document to the DLQ", e);
-          dlqObject.releaseEventHandle(false);
-        }
-      });
-    } else if (dlqWriter != null) {
-      try {
-        dlqWriter.write(dlqObjects, pipeline, PLUGIN_NAME);
-        dlqObjects.forEach((dlqObject) -> {
-          dlqObject.releaseEventHandle(true);
-        });
-      } catch (final IOException e) {
-        dlqObjects.forEach(dlqObject -> {
-          LOG.error(NOISY, "Failed to write a document to the DLQ", e);
-          dlqObject.releaseEventHandle(false);
-        });
-      }
-    } else {
-      dlqObjects.forEach(dlqObject -> {
-
-        final FailedDlqData failedDlqData = (FailedDlqData) dlqObject.getFailedData();
-
-        final String message = failure == null ? failedDlqData.getMessage() : failure.getMessage();
-        LOG.warn("Document failed to write to OpenSearch with error code {}. Configure a DLQ to save failed documents. Error: {}", failedDlqData.getStatus(), message);
-        dlqObject.releaseEventHandle(false);
-      });
-    }
-  }
-
-  private void closeFiles() {
-    // Close the client. This closes the low-level client which will close it for both high-level clients.
+  @Override
+  public void shutdown() {
+    super.shutdown();
+    ingester.shutdown();
     if (restHighLevelClient != null) {
       try {
         restHighLevelClient.close();
@@ -649,32 +189,8 @@ public class OpenSearchSink extends AbstractSink<Record<Event>> {
         throw new RuntimeException(e.getMessage(), e);
       }
     }
-    if (dlqWriter != null) {
-      try {
-        dlqWriter.close();
-      } catch (final IOException e) {
-        LOG.error(e.getMessage(), e);
-      }
-    }
-    if (dlqFileWriter != null) {
-      try {
-        dlqFileWriter.close();
-      } catch (final IOException e) {
-        LOG.error(e.getMessage(), e);
-      }
-    }
-  }
-
-  @Override
-  public void shutdown() {
-    super.shutdown();
-    closeFiles();
     if (openSearchClient != null) {
       openSearchClient.shutdown();
-    }
-    if (queryExecutorService != null && existingDocumentQueryManager != null) {
-      existingDocumentQueryManager.stop();
-      queryExecutorService.shutdown();
     }
   }
 
@@ -692,75 +208,5 @@ public class OpenSearchSink extends AbstractSink<Record<Event>> {
           maybeServerlessOptions.get().getVpceId()
       );
     }
-  }
-
-  private static boolean isExternalVersionType(final VersionType versionType) {
-    return versionType != null && (versionType == VersionType.External || versionType == VersionType.ExternalGte);
-  }
-
-  private DlqObject createDlqObjectFromEvent(final Event event,
-                                             final String index,
-                                             final String message) {
-    DlqObject.Builder builder = DlqObject.builder()
-            .withFailedData(FailedDlqData.builder()
-                    .withDocument(event.toJsonString())
-                    .withIndex(index)
-                    .withMessage(message != null ? message : "")
-                    .build())
-            .withPluginName(PLUGIN_NAME)
-            .withPipelineName(pipeline)
-            .withPluginId(PLUGIN_NAME);
-
-    if (useEventInBulkOperation) {
-      builder.withEvent(event);
-    } else {
-      builder.withEventHandle(event.getEventHandle());
-    }
-    return builder.build();
-  }
-
-  private String resolveEventAction(final Event event) {
-    String resolvedAction = action;
-    if (actions != null) {
-      for (final ActionConfiguration actionEntry : actions) {
-        final String condition = actionEntry.getWhen();
-        resolvedAction = actionEntry.getType();
-        if (condition != null &&
-                expressionEvaluator.evaluateConditional(condition, event)) {
-          break;
-        }
-      }
-    }
-    if (resolvedAction.contains("${")) {
-      resolvedAction = event.formatString(resolvedAction, expressionEvaluator);
-    }
-    return resolvedAction;
-  }
-
-  /**
-   * This function is used for update and upsert bulk actions to determine whether the original JsonNode needs to be filtered down
-   * based on the user's sink configuration. If a new parameter manipulates the document before sending to OpenSearch, it needs to be added to
-   * this list to get applied for update and upsert actions
-   * @return whether the doc
-   */
-  private boolean isUsingDocumentFilters() {
-    return documentRootKey != null ||
-            (sinkContext.getIncludeKeys() != null && !sinkContext.getIncludeKeys().isEmpty()) ||
-            (sinkContext.getExcludeKeys() != null && !sinkContext.getExcludeKeys().isEmpty()) ||
-            sinkContext.getTagsTargetKey() != null;
-  }
-
-  private AccumulatingBulkRequest<BulkOperationWrapper, BulkRequest> flushBatch(
-          final AccumulatingBulkRequest<BulkOperationWrapper, BulkRequest> bulkRequest,
-          final BulkOperationWrapper bulkOperationWrapper,
-          long lastFlushTime
-          ) {
-    final long estimatedBytesBeforeAdd = bulkRequest.estimateSizeInBytesWithDocument(bulkOperationWrapper);
-    if (bulkSize >= 0 && estimatedBytesBeforeAdd >= bulkSize && bulkRequest.getOperationsCount() > 0) {
-      flushBatch(bulkRequest);
-      lastFlushTime = System.currentTimeMillis();
-      return bulkRequestSupplier.get();
-    }
-   return bulkRequest;
   }
 }
