+++ data-prepper-plugins/opensearch/src/test/java/org/opensearch/dataprepper/plugins/sink/opensearch/OpenSearchSinkTest.java
@@ -1,6 +1,10 @@
 /*
  * Copyright OpenSearch Contributors
  * SPDX-License-Identifier: Apache-2.0
+ *
+ * The OpenSearch Contributors require contributions made to
+ * this file be licensed under the Apache-2.0 license or a
+ * compatible open source license.
  */
 
 package org.opensearch.dataprepper.plugins.sink.opensearch;
@@ -11,59 +15,37 @@ import io.micrometer.core.instrument.Timer;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
-import org.junit.jupiter.params.provider.MethodSource;
-import org.mockito.ArgumentCaptor;
 import org.mockito.Mock;
 import org.mockito.MockedConstruction;
 import org.mockito.MockedStatic;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.opensearch.client.RestHighLevelClient;
 import org.opensearch.client.opensearch.OpenSearchClient;
-import org.opensearch.client.opensearch._types.VersionType;
 import org.opensearch.dataprepper.aws.api.AwsCredentialsSupplier;
 import org.opensearch.dataprepper.expression.ExpressionEvaluator;
-import org.opensearch.dataprepper.metrics.MetricNames;
 import org.opensearch.dataprepper.metrics.PluginMetrics;
 import org.opensearch.dataprepper.model.configuration.PipelineDescription;
 import org.opensearch.dataprepper.model.configuration.PluginSetting;
-import org.opensearch.dataprepper.model.pipeline.HeadlessPipeline;
-import org.opensearch.dataprepper.model.sink.SinkForwardRecordsContext;
 import org.opensearch.dataprepper.model.event.Event;
-import org.opensearch.dataprepper.model.event.EventHandle;
-import org.opensearch.dataprepper.model.event.JacksonEvent;
-import org.opensearch.dataprepper.model.failures.DlqObject;
 import org.opensearch.dataprepper.model.plugin.PluginConfigObservable;
 import org.opensearch.dataprepper.model.record.Record;
 import org.opensearch.dataprepper.model.sink.SinkContext;
 import org.opensearch.dataprepper.plugins.sink.opensearch.configuration.OpenSearchSinkConfig;
-import org.opensearch.dataprepper.plugins.sink.opensearch.dlq.FailedDlqData;
-import org.opensearch.dataprepper.plugins.sink.opensearch.index.DocumentBuilder;
 import org.opensearch.dataprepper.plugins.sink.opensearch.index.IndexConfiguration;
 import org.opensearch.dataprepper.plugins.sink.opensearch.index.IndexManager;
 import org.opensearch.dataprepper.plugins.sink.opensearch.index.IndexManagerFactory;
-import org.opensearch.dataprepper.plugins.sink.opensearch.index.DataStreamDetector;
-import org.opensearch.dataprepper.plugins.sink.opensearch.index.DataStreamIndex;
-import org.opensearch.dataprepper.plugins.sink.opensearch.index.IndexCache;
-import org.opensearch.dataprepper.test.helper.ReflectivelySetField;
 import org.opensearch.dataprepper.plugins.sink.opensearch.index.IndexType;
 import org.opensearch.dataprepper.plugins.sink.opensearch.index.TemplateStrategy;
 import org.opensearch.dataprepper.plugins.sink.opensearch.index.TemplateType;
 
 import java.io.IOException;
-import java.lang.reflect.Field;
 import java.util.Collections;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
-import java.util.stream.Stream;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
-import static org.hamcrest.Matchers.notNullValue;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.ArgumentMatchers.eq;
@@ -77,6 +59,8 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import static org.opensearch.dataprepper.model.sink.SinkLatencyMetrics.EXTERNAL_LATENCY;
 import static org.opensearch.dataprepper.model.sink.SinkLatencyMetrics.INTERNAL_LATENCY;
+import static org.opensearch.dataprepper.metrics.MetricNames.RECORDS_IN;
+import static org.opensearch.dataprepper.metrics.MetricNames.TIME_ELAPSED;
 import static org.opensearch.dataprepper.plugins.sink.opensearch.OpenSearchSink.BULKREQUEST_ERRORS;
 import static org.opensearch.dataprepper.plugins.sink.opensearch.OpenSearchSink.BULKREQUEST_LATENCY;
 import static org.opensearch.dataprepper.plugins.sink.opensearch.OpenSearchSink.BULKREQUEST_SIZE_BYTES;
@@ -123,44 +107,17 @@ public class OpenSearchSinkTest {
     private PluginMetrics pluginMetrics;
 
     @Mock
-    private Timer bulkRequestTimer;
-
-    @Mock
     private OpenSearchSinkConfig openSearchSinkConfig;
 
     @Mock
-    private Counter bulkRequestErrorsCounter;
-
-    @Mock
-    private Counter invalidActionErrorsCounter;
-
-    @Mock
-    private Counter dynamicIndexDroppedEvents;
-
-    @Mock
-    private DistributionSummary bulkRequestSizeBytesSummary;
-
-    @Mock
-    private Counter dynamicDocumentVersionDroppedEvents;
-
-    @Mock
     private PluginConfigObservable pluginConfigObservable;
-
-    @Mock
-    private DataStreamDetector dataStreamDetector;
-    
-    @Mock
-    private DataStreamIndex dataStreamIndex;
-    
-    @Mock
-    private IndexCache indexCache;
 
     @BeforeEach
     void setup() {
         when(pipelineDescription.getPipelineName()).thenReturn(UUID.randomUUID().toString());
 
         final RetryConfiguration retryConfiguration = mock(RetryConfiguration.class);
-        when(retryConfiguration.getDlq()).thenReturn(Optional.empty());
+        lenient().when(retryConfiguration.getDlq()).thenReturn(Optional.empty());
         lenient().when(retryConfiguration.getDlqFile()).thenReturn(null);
 
         final ConnectionConfiguration connectionConfiguration = mock(ConnectionConfiguration.class);
@@ -168,34 +125,35 @@ public class OpenSearchSinkTest {
         lenient().when(connectionConfiguration.createClient(awsCredentialsSupplier)).thenReturn(restHighLevelClient);
         lenient().when(connectionConfiguration.createOpenSearchClient(restHighLevelClient, awsCredentialsSupplier)).thenReturn(openSearchClient);
 
-        when(indexConfiguration.getAction()).thenReturn("index");
-        when(indexConfiguration.getDocumentId()).thenReturn(null);
-        when(indexConfiguration.getDocumentIdField()).thenReturn(null);
-        when(indexConfiguration.getRouting()).thenReturn(null);
-        when(indexConfiguration.getActions()).thenReturn(null);
-        when(indexConfiguration.getDocumentRootKey()).thenReturn(null);
+        lenient().when(indexConfiguration.getAction()).thenReturn("index");
+        lenient().when(indexConfiguration.getDocumentId()).thenReturn(null);
+        lenient().when(indexConfiguration.getDocumentIdField()).thenReturn(null);
+        lenient().when(indexConfiguration.getRouting()).thenReturn(null);
+        lenient().when(indexConfiguration.getActions()).thenReturn(null);
+        lenient().when(indexConfiguration.getDocumentRootKey()).thenReturn(null);
         lenient().when(indexConfiguration.getVersionType()).thenReturn(null);
         lenient().when(indexConfiguration.getVersionExpression()).thenReturn(null);
         lenient().when(indexConfiguration.getIndexAlias()).thenReturn(UUID.randomUUID().toString());
         lenient().when(indexConfiguration.getTemplateType()).thenReturn(TemplateType.V1);
-        when(indexConfiguration.getIndexType()).thenReturn(IndexType.CUSTOM);
-        when(indexConfiguration.getBulkSize()).thenReturn(DEFAULT_BULK_SIZE);
-        when(indexConfiguration.getFlushTimeout()).thenReturn(DEFAULT_FLUSH_TIMEOUT);
+        lenient().when(indexConfiguration.getIndexType()).thenReturn(IndexType.CUSTOM);
+        lenient().when(indexConfiguration.getBulkSize()).thenReturn(DEFAULT_BULK_SIZE);
+        lenient().when(indexConfiguration.getFlushTimeout()).thenReturn(DEFAULT_FLUSH_TIMEOUT);
 
-        when(openSearchSinkConfiguration.getIndexConfiguration()).thenReturn(indexConfiguration);
-        when(openSearchSinkConfiguration.getRetryConfiguration()).thenReturn(retryConfiguration);
+        lenient().when(openSearchSinkConfiguration.getIndexConfiguration()).thenReturn(indexConfiguration);
+        lenient().when(openSearchSinkConfiguration.getRetryConfiguration()).thenReturn(retryConfiguration);
         lenient().when(openSearchSinkConfiguration.getConnectionConfiguration()).thenReturn(connectionConfiguration);
 
-        when(pluginMetrics.counter(MetricNames.RECORDS_IN)).thenReturn(mock(Counter.class));
-        when(pluginMetrics.timer(MetricNames.TIME_ELAPSED)).thenReturn(mock(Timer.class));
-        when(pluginMetrics.timer(INTERNAL_LATENCY)).thenReturn(mock(Timer.class));
-        when(pluginMetrics.timer(EXTERNAL_LATENCY)).thenReturn(mock(Timer.class));
-        when(pluginMetrics.timer(BULKREQUEST_LATENCY)).thenReturn(bulkRequestTimer);
-        when(pluginMetrics.counter(BULKREQUEST_ERRORS)).thenReturn(bulkRequestErrorsCounter);
-        when(pluginMetrics.counter(INVALID_ACTION_ERRORS)).thenReturn(invalidActionErrorsCounter);
-        when(pluginMetrics.counter(DYNAMIC_INDEX_DROPPED_EVENTS)).thenReturn(dynamicIndexDroppedEvents);
-        when(pluginMetrics.counter(INVALID_VERSION_EXPRESSION_DROPPED_EVENTS)).thenReturn(dynamicDocumentVersionDroppedEvents);
-        when(pluginMetrics.summary(BULKREQUEST_SIZE_BYTES)).thenReturn(bulkRequestSizeBytesSummary);
+        lenient().when(pluginMetrics.counter(RECORDS_IN)).thenReturn(mock(Counter.class));
+        lenient().when(pluginMetrics.timer(TIME_ELAPSED)).thenReturn(mock(Timer.class));
+        lenient().when(pluginMetrics.timer(INTERNAL_LATENCY)).thenReturn(mock(Timer.class));
+        lenient().when(pluginMetrics.timer(EXTERNAL_LATENCY)).thenReturn(mock(Timer.class));
+        lenient().when(pluginMetrics.timer(BULKREQUEST_LATENCY)).thenReturn(mock(Timer.class));
+        lenient().when(pluginMetrics.counter(BULKREQUEST_ERRORS)).thenReturn(mock(Counter.class));
+        lenient().when(pluginMetrics.counter(INVALID_ACTION_ERRORS)).thenReturn(mock(Counter.class));
+        lenient().when(pluginMetrics.counter(DYNAMIC_INDEX_DROPPED_EVENTS)).thenReturn(mock(Counter.class));
+        lenient().when(pluginMetrics.counter(INVALID_VERSION_EXPRESSION_DROPPED_EVENTS)).thenReturn(mock(Counter.class));
+        lenient().when(pluginMetrics.summary(BULKREQUEST_SIZE_BYTES)).thenReturn(mock(DistributionSummary.class));
+        lenient().when(pluginMetrics.counter(anyString())).thenReturn(mock(Counter.class));
 
         lenient().when(sinkContext.getTagsTargetKey()).thenReturn(null);
         lenient().when(sinkContext.getIncludeKeys()).thenReturn(null);
@@ -207,7 +165,8 @@ public class OpenSearchSinkTest {
              final MockedStatic<PluginMetrics> pluginMetricsMockedStatic = mockStatic(PluginMetrics.class);
              final MockedConstruction<IndexManagerFactory> indexManagerFactoryMockedConstruction = mockConstruction(IndexManagerFactory.class, (mock, context) -> {
                  indexManagerFactory = mock;
-             })) {
+             });
+             final MockedConstruction<BulkIngester> bulkIngesterMockedConstruction = mockConstruction(BulkIngester.class)) {
             pluginMetricsMockedStatic.when(() -> PluginMetrics.fromPluginSetting(pluginSetting)).thenReturn(pluginMetrics);
             openSearchSinkConfigurationMockedStatic.when(() -> OpenSearchSinkConfiguration.readOSConfig(openSearchSinkConfig, expressionEvaluator))
                     .thenReturn(openSearchSinkConfiguration);
@@ -225,12 +184,11 @@ public class OpenSearchSinkTest {
         objectUnderTest.initialize();
 
         verify(pluginConfigObservable).addPluginConfigObserver(any());
+        assertThat(objectUnderTest.isReady(), equalTo(true));
     }
 
     @Test
-    void test_initialization_with_failure_and_retry_with_query_manager() throws IOException {
-        when(openSearchSinkConfiguration.getIndexConfiguration().getQueryTerm()).thenReturn(UUID.randomUUID().toString());
-
+    void test_initialization_with_failure_and_retry() throws IOException {
         final OpenSearchSink objectUnderTest = createObjectUnderTest();
         when(indexManagerFactory.getIndexManager(any(IndexType.class), eq(openSearchClient), any(RestHighLevelClient.class), eq(openSearchSinkConfiguration), any(TemplateStrategy.class), any()))
                 .thenThrow(RuntimeException.class).thenReturn(indexManager);
@@ -241,423 +199,29 @@ public class OpenSearchSinkTest {
     }
 
     @Test
-    void test_sink_successful_records_handling_without_forwarding_pipelines_bulk_operations_with_event_handles() throws Exception {
-        when(sinkContext.getForwardToPipelines()).thenReturn(Map.of());
+    void doOutput_delegates_to_ingester() throws Exception {
         final OpenSearchSink objectUnderTest = createObjectUnderTest();
-        BulkOperationWrapper op1 = mock(BulkOperationWrapper.class);
-        BulkOperationWrapper op2 = mock(BulkOperationWrapper.class);
-        when(op1.getEvent()).thenReturn(null);
-        when(op2.getEvent()).thenReturn(null);
-        List<BulkOperationWrapper> operationsList = new ArrayList<>();
-        operationsList.add(op1);
-        operationsList.add(op2);
-        objectUnderTest.successfulOperationsHandler(operationsList);
-        verify(op1).releaseEventHandle(eq(true));
-        verify(op2).releaseEventHandle(eq(true));
 
+        final Ingester ingester = getField(objectUnderTest, "ingester");
+        final List<Record<Event>> records = Collections.emptyList();
+        objectUnderTest.doOutput(records);
+
+        verify(ingester).output(records);
     }
 
     @Test
-    void test_sink_successful_records_handling_with_forwarding_pipelines() throws Exception {
-        HeadlessPipeline forwardPipeline = mock(HeadlessPipeline.class);
-        when(sinkContext.getForwardToPipelines()).thenReturn(Map.of("fwd_pipeline", forwardPipeline));
-        final EventHandle eventHandle = mock(EventHandle.class);
+    void shutdown_delegates_to_ingester() throws Exception {
         final OpenSearchSink objectUnderTest = createObjectUnderTest();
-        BulkOperationWrapper op1 = mock(BulkOperationWrapper.class);
-        BulkOperationWrapper op2 = mock(BulkOperationWrapper.class);
-        Event event = mock(Event.class);
-        when(op1.getEvent()).thenReturn(event);
-        when(op2.getEvent()).thenReturn(event);
-        List<BulkOperationWrapper> operationsList = new ArrayList<>();
-        operationsList.add(op1);
-        operationsList.add(op2);
-        objectUnderTest.successfulOperationsHandler(operationsList);
-        verify(eventHandle, times(0)).release(eq(true));
-        verify(op1, times(1)).getEvent();
-        verify(op2, times(1)).getEvent();
-        verify(sinkContext, times(1)).forwardRecords(any(SinkForwardRecordsContext.class), eq(null), eq(null));
-        
-    }
 
- 
-    @Test
-    void test_sink_successful_records_handling_without_forwarding_pipelines_bulk_operations_with_events() throws Exception {
-        when(sinkContext.getForwardToPipelines()).thenReturn(Map.of());
-        final EventHandle eventHandle = mock(EventHandle.class);
-        final OpenSearchSink objectUnderTest = createObjectUnderTest();
-        BulkOperationWrapper op1 = mock(BulkOperationWrapper.class);
-        BulkOperationWrapper op2 = mock(BulkOperationWrapper.class);
-        Event event = mock(Event.class);
-        when(event.getEventHandle()).thenReturn(eventHandle);
-        when(op1.getEvent()).thenReturn(event);
-        when(op2.getEvent()).thenReturn(event);
-        List<BulkOperationWrapper> operationsList = new ArrayList<>();
-        operationsList.add(op1);
-        operationsList.add(op2);
-        objectUnderTest.successfulOperationsHandler(operationsList);
-        verify(eventHandle, times(2)).release(eq(true));
-        verify(op1, times(0)).getEventHandle();
-        verify(op2, times(0)).getEventHandle();
-        
-    }
+        final Ingester ingester = getField(objectUnderTest, "ingester");
+        objectUnderTest.shutdown();
 
-    @Test
-    void doOutput_with_invalid_version_expression_catches_NumberFormatException_and_creates_DLQObject() throws IOException {
-        when(pluginSetting.getName()).thenReturn("opensearch");
-        final String versionExpression = UUID.randomUUID().toString();
-        when(indexConfiguration.getVersionExpression()).thenReturn(versionExpression);
-
-        final Event event = mock(JacksonEvent.class);
-        final String document = UUID.randomUUID().toString();
-        when(event.toJsonString()).thenReturn(document);
-        final EventHandle eventHandle = mock(EventHandle.class);
-        when(event.getEventHandle()).thenReturn(eventHandle);
-        final String index = UUID.randomUUID().toString();
-        when(event.formatString(versionExpression, expressionEvaluator)).thenReturn("not_a_number");
-        when(event.formatString(indexConfiguration.getIndexAlias(), expressionEvaluator)).thenReturn(index);
-        final Record<Event> eventRecord = new Record<>(event);
-
-        final OpenSearchSink objectUnderTest = createObjectUnderTest();
-        when(indexManagerFactory.getIndexManager(any(IndexType.class), eq(openSearchClient), any(RestHighLevelClient.class), eq(openSearchSinkConfiguration), any(TemplateStrategy.class), any()))
-                .thenReturn(indexManager);
-        doNothing().when(indexManager).setupIndex();
-        objectUnderTest.initialize();
-
-        when(indexManager.getIndexName(anyString())).thenReturn(index);
-
-        final DlqObject dlqObject = mock(DlqObject.class);
-
-        final DlqObject.Builder dlqObjectBuilder = mock(DlqObject.Builder.class);
-        final ArgumentCaptor<FailedDlqData> failedDlqData = ArgumentCaptor.forClass(FailedDlqData.class);
-        when(dlqObjectBuilder.withEventHandle(eventHandle)).thenReturn(dlqObjectBuilder);
-        when(dlqObjectBuilder.withFailedData(failedDlqData.capture())).thenReturn(dlqObjectBuilder);
-        when(dlqObjectBuilder.withPluginName(pluginSetting.getName())).thenReturn(dlqObjectBuilder);
-        when(dlqObjectBuilder.withPluginId(pluginSetting.getName())).thenReturn(dlqObjectBuilder);
-        when(dlqObjectBuilder.withPipelineName(pipelineDescription.getPipelineName())).thenReturn(dlqObjectBuilder);
-
-        when(dlqObject.getFailedData()).thenReturn(mock(FailedDlqData.class));
-        doNothing().when(dlqObject).releaseEventHandle(false);
-        when(dlqObjectBuilder.build()).thenReturn(dlqObject);
-
-        try (final MockedStatic<DocumentBuilder> documentBuilderMockedStatic = mockStatic(DocumentBuilder.class);
-             final MockedStatic<DlqObject> dlqObjectMockedStatic = mockStatic(DlqObject.class)) {
-            documentBuilderMockedStatic.when(() -> DocumentBuilder.build(eq(event), eq(null), eq(null), eq(null), eq(null)))
-                    .thenReturn(UUID.randomUUID().toString());
-
-            dlqObjectMockedStatic.when(DlqObject::builder).thenReturn(dlqObjectBuilder);
-            objectUnderTest.doOutput(List.of(eventRecord));
-        }
-
-        final FailedDlqData failedDlqDataResult = failedDlqData.getValue();
-        assertThat(failedDlqDataResult, notNullValue());
-        assertThat(failedDlqDataResult.getDocument(), equalTo(document));
-        assertThat(failedDlqDataResult.getIndex(), equalTo(index));
-        assertThat(failedDlqDataResult.getMessage().startsWith("Unable to convert the result of evaluating document_version"), equalTo(true));
-
-        verify(dynamicDocumentVersionDroppedEvents).increment();
-    }
-
-    @ParameterizedTest
-    @MethodSource("invalidVersionExceptionProvider")
-    void doOutput_with_invalid_version_expression_does_not_add_event_to_bulk_request(
-            final Class<? extends RuntimeException> exceptionType) throws IOException {
-        when(pluginSetting.getName()).thenReturn("opensearch");
-        final String versionExpression = UUID.randomUUID().toString();
-        when(indexConfiguration.getVersionExpression()).thenReturn(versionExpression);
-
-        final Event event = mock(JacksonEvent.class);
-        final String document = UUID.randomUUID().toString();
-        when(event.toJsonString()).thenReturn(document);
-        final EventHandle eventHandle = mock(EventHandle.class);
-        when(event.getEventHandle()).thenReturn(eventHandle);
-        final String index = UUID.randomUUID().toString();
-        when(event.formatString(versionExpression, expressionEvaluator)).thenThrow(exceptionType);
-        when(event.formatString(indexConfiguration.getIndexAlias(), expressionEvaluator)).thenReturn(index);
-        final Record<Event> eventRecord = new Record<>(event);
-
-        final OpenSearchSink objectUnderTest = createObjectUnderTest();
-        when(indexManagerFactory.getIndexManager(any(IndexType.class), eq(openSearchClient), any(RestHighLevelClient.class), eq(openSearchSinkConfiguration), any(TemplateStrategy.class), any()))
-                .thenReturn(indexManager);
-        doNothing().when(indexManager).setupIndex();
-        objectUnderTest.initialize();
-
-        when(indexManager.getIndexName(anyString())).thenReturn(index);
-
-        final DlqObject dlqObject = mock(DlqObject.class);
-        final DlqObject.Builder dlqObjectBuilder = mock(DlqObject.Builder.class);
-        when(dlqObjectBuilder.withEventHandle(eventHandle)).thenReturn(dlqObjectBuilder);
-        when(dlqObjectBuilder.withFailedData(any(FailedDlqData.class))).thenReturn(dlqObjectBuilder);
-        when(dlqObjectBuilder.withPluginName(pluginSetting.getName())).thenReturn(dlqObjectBuilder);
-        when(dlqObjectBuilder.withPluginId(pluginSetting.getName())).thenReturn(dlqObjectBuilder);
-        when(dlqObjectBuilder.withPipelineName(pipelineDescription.getPipelineName())).thenReturn(dlqObjectBuilder);
-        when(dlqObject.getFailedData()).thenReturn(mock(FailedDlqData.class));
-        doNothing().when(dlqObject).releaseEventHandle(false);
-        when(dlqObjectBuilder.build()).thenReturn(dlqObject);
-
-        try (final MockedStatic<DocumentBuilder> documentBuilderMockedStatic = mockStatic(DocumentBuilder.class);
-             final MockedStatic<DlqObject> dlqObjectMockedStatic = mockStatic(DlqObject.class)) {
-            documentBuilderMockedStatic.when(() -> DocumentBuilder.build(eq(event), eq(null), eq(null), eq(null), eq(null)))
-                    .thenReturn(UUID.randomUUID().toString());
-            dlqObjectMockedStatic.when(DlqObject::builder).thenReturn(dlqObjectBuilder);
-            objectUnderTest.doOutput(List.of(eventRecord));
-        }
-
-        verify(dynamicDocumentVersionDroppedEvents).increment();
-        verify(event, times(0)).getJsonNode();
-    }
-
-    private static Stream<Arguments> invalidVersionExceptionProvider() {
-        return Stream.of(
-                Arguments.of(NumberFormatException.class),
-                Arguments.of(RuntimeException.class)
-        );
-    }
-
-    @Test
-    void test_routing_field_in_document() throws IOException {
-        String routingFieldKey = UUID.randomUUID().toString();
-        String routingKey = UUID.randomUUID().toString();
-        String routingFieldValue = UUID.randomUUID().toString();
-        when(indexConfiguration.getRoutingField()).thenReturn(routingFieldKey);
-        when(indexConfiguration.getRouting()).thenReturn(routingKey);
-        final OpenSearchSink objectUnderTest = createObjectUnderTest();
-        final Event event = JacksonEvent.builder()
-                .withEventType("event")
-                .withData(Collections.singletonMap(routingFieldKey, routingFieldValue))
-                .build();
-        assertThat(objectUnderTest.getDocument(event).getRoutingField(), equalTo(Optional.of(routingFieldValue)));
-    }
-
-    @Test
-    void test_routing_in_document() throws IOException {
-        String routingValue = UUID.randomUUID().toString();
-        String routingKey = UUID.randomUUID().toString();
-        final OpenSearchSink objectUnderTest = createObjectUnderTest();
-        final Event event = JacksonEvent.builder()
-                .withEventType("event")
-                .withData(Collections.singletonMap(routingKey, routingValue))
-                .build();
-        assertThat(objectUnderTest.getDocument(event).getRoutingField(), equalTo(Optional.empty()));
-
-        when(indexConfiguration.getRouting()).thenReturn("${"+routingKey+"}");
-        final OpenSearchSink objectUnderTest2 = createObjectUnderTest();
-        assertThat(objectUnderTest2.getDocument(event).getRoutingField(), equalTo(Optional.of(routingValue)));
-    }
-
-    @Test
-    void doOutput_with_invalid_version_expression_result_catches_RuntimeException_and_creates_DLQObject() throws IOException {
-
-        when(pluginSetting.getName()).thenReturn("opensearch");
-        final String versionExpression = UUID.randomUUID().toString();
-        when(indexConfiguration.getVersionExpression()).thenReturn(versionExpression);
-
-        final Event event = mock(JacksonEvent.class);
-        final String document = UUID.randomUUID().toString();
-        when(event.toJsonString()).thenReturn(document);
-        final EventHandle eventHandle = mock(EventHandle.class);
-        when(event.getEventHandle()).thenReturn(eventHandle);
-        final String index = UUID.randomUUID().toString();
-        when(event.formatString(versionExpression, expressionEvaluator)).thenThrow(RuntimeException.class);
-        when(event.formatString(indexConfiguration.getIndexAlias(), expressionEvaluator)).thenReturn(index);
-        final Record<Event> eventRecord = new Record<>(event);
-
-        final OpenSearchSink objectUnderTest = createObjectUnderTest();
-        when(indexManagerFactory.getIndexManager(any(IndexType.class), eq(openSearchClient), any(RestHighLevelClient.class), eq(openSearchSinkConfiguration), any(TemplateStrategy.class), any()))
-                .thenReturn(indexManager);
-        doNothing().when(indexManager).setupIndex();
-        objectUnderTest.initialize();
-
-        when(indexManager.getIndexName(anyString())).thenReturn(index);
-
-        final DlqObject dlqObject = mock(DlqObject.class);
-
-        final DlqObject.Builder dlqObjectBuilder = mock(DlqObject.Builder.class);
-        final ArgumentCaptor<FailedDlqData> failedDlqData = ArgumentCaptor.forClass(FailedDlqData.class);
-        when(dlqObjectBuilder.withEventHandle(eventHandle)).thenReturn(dlqObjectBuilder);
-        when(dlqObjectBuilder.withFailedData(failedDlqData.capture())).thenReturn(dlqObjectBuilder);
-        when(dlqObjectBuilder.withPluginName(pluginSetting.getName())).thenReturn(dlqObjectBuilder);
-        when(dlqObjectBuilder.withPluginId(pluginSetting.getName())).thenReturn(dlqObjectBuilder);
-        when(dlqObjectBuilder.withPipelineName(pipelineDescription.getPipelineName())).thenReturn(dlqObjectBuilder);
-
-        when(dlqObject.getFailedData()).thenReturn(mock(FailedDlqData.class));
-        doNothing().when(dlqObject).releaseEventHandle(false);
-        when(dlqObjectBuilder.build()).thenReturn(dlqObject);
-
-        try (final MockedStatic<DocumentBuilder> documentBuilderMockedStatic = mockStatic(DocumentBuilder.class);
-             final MockedStatic<DlqObject> dlqObjectMockedStatic = mockStatic(DlqObject.class)) {
-            documentBuilderMockedStatic.when(() -> DocumentBuilder.build(eq(event), eq(null), eq(null), eq(null), eq(null)))
-                    .thenReturn(UUID.randomUUID().toString());
-
-            dlqObjectMockedStatic.when(DlqObject::builder).thenReturn(dlqObjectBuilder);
-            objectUnderTest.doOutput(List.of(eventRecord));
-        }
-
-        final FailedDlqData failedDlqDataResult = failedDlqData.getValue();
-        assertThat(failedDlqDataResult, notNullValue());
-        assertThat(failedDlqDataResult.getDocument(), equalTo(document));
-        assertThat(failedDlqDataResult.getIndex(), equalTo(index));
-        assertThat(failedDlqDataResult.getMessage().startsWith("There was an exception when evaluating the document_version"), equalTo(true));
-
-        verify(dynamicDocumentVersionDroppedEvents).increment();
-    }
-
-    @Test
-    void createDlqObjectFromEvent_with_null_message_uses_default_message() throws Exception {
-        when(pluginSetting.getName()).thenReturn("opensearch");
-        
-        final Event event = mock(JacksonEvent.class);
-        final String document = UUID.randomUUID().toString();
-        when(event.toJsonString()).thenReturn(document);
-        final EventHandle eventHandle = mock(EventHandle.class);
-        when(event.getEventHandle()).thenReturn(eventHandle);
-        final String index = UUID.randomUUID().toString();
-        
-        final OpenSearchSink objectUnderTest = createObjectUnderTest();
-        when(indexManagerFactory.getIndexManager(any(IndexType.class), eq(openSearchClient), any(RestHighLevelClient.class), eq(openSearchSinkConfiguration), any(TemplateStrategy.class), any()))
-                .thenReturn(indexManager);
-        doNothing().when(indexManager).setupIndex();
-        objectUnderTest.initialize();
-        
-        final DlqObject.Builder dlqObjectBuilder = mock(DlqObject.Builder.class);
-        final ArgumentCaptor<FailedDlqData> failedDlqData = ArgumentCaptor.forClass(FailedDlqData.class);
-        when(dlqObjectBuilder.withEventHandle(eventHandle)).thenReturn(dlqObjectBuilder);
-        when(dlqObjectBuilder.withFailedData(failedDlqData.capture())).thenReturn(dlqObjectBuilder);
-        when(dlqObjectBuilder.withPluginName(pluginSetting.getName())).thenReturn(dlqObjectBuilder);
-        when(dlqObjectBuilder.withPluginId(pluginSetting.getName())).thenReturn(dlqObjectBuilder);
-        when(dlqObjectBuilder.withPipelineName(pipelineDescription.getPipelineName())).thenReturn(dlqObjectBuilder);
-        when(dlqObjectBuilder.build()).thenReturn(mock(DlqObject.class));
-        
-        try (final MockedStatic<DlqObject> dlqObjectMockedStatic = mockStatic(DlqObject.class)) {
-            dlqObjectMockedStatic.when(DlqObject::builder).thenReturn(dlqObjectBuilder);
-            
-            // Use reflection to call the private method with null message
-            java.lang.reflect.Method method = OpenSearchSink.class.getDeclaredMethod("createDlqObjectFromEvent", Event.class, String.class, String.class);
-            method.setAccessible(true);
-            method.invoke(objectUnderTest, event, index, null);
-        }
-        
-        final FailedDlqData failedDlqDataResult = failedDlqData.getValue();
-        assertThat(failedDlqDataResult, notNullValue());
-        assertThat(failedDlqDataResult.getDocument(), equalTo(document));
-        assertThat(failedDlqDataResult.getIndex(), equalTo(index));
-        assertThat(failedDlqDataResult.getMessage(), equalTo(""));
-    }
-
-    @Test
-    void dataStreamIndex_determineAction_returnsCreate_whenIndexIsDataStream() throws Exception {
-        when(indexConfiguration.getAction()).thenReturn("index");
-        
-        final OpenSearchSink objectUnderTest = createObjectUnderTest();
-        when(indexManagerFactory.getIndexManager(any(IndexType.class), eq(openSearchClient), any(RestHighLevelClient.class), eq(openSearchSinkConfiguration), any(TemplateStrategy.class), any()))
-                .thenReturn(indexManager);
-        doNothing().when(indexManager).setupIndex();
-        objectUnderTest.initialize();
-        
-        // Use ReflectivelySetField to set the dataStreamIndex
-        ReflectivelySetField.setField(OpenSearchSink.class, objectUnderTest, "dataStreamIndex", dataStreamIndex);
-        
-        // Verify the dataStreamIndex was set correctly
-        assertThat(objectUnderTest, notNullValue());
-    }
-
-    @Test
-    void dataStreamIndex_determineAction_returnsConfiguredAction_whenIndexIsNotDataStream() throws Exception {
-        when(indexConfiguration.getAction()).thenReturn("index");
-        
-        final OpenSearchSink objectUnderTest = createObjectUnderTest();
-        when(indexManagerFactory.getIndexManager(any(IndexType.class), eq(openSearchClient), any(RestHighLevelClient.class), eq(openSearchSinkConfiguration), any(TemplateStrategy.class), any()))
-                .thenReturn(indexManager);
-        doNothing().when(indexManager).setupIndex();
-        objectUnderTest.initialize();
-        
-        // Use ReflectivelySetField to set the dataStreamIndex
-        ReflectivelySetField.setField(OpenSearchSink.class, objectUnderTest, "dataStreamIndex", dataStreamIndex);
-        
-        // Verify the dataStreamIndex was set correctly
-        assertThat(objectUnderTest, notNullValue());
-    }
-    
-    @Test
-    void dataStreamIndex_ensureTimestamp_doesNotOverwriteExistingTimestamp() throws Exception {
-        final OpenSearchSink objectUnderTest = createObjectUnderTest();
-        when(indexManagerFactory.getIndexManager(any(IndexType.class), eq(openSearchClient), any(RestHighLevelClient.class), eq(openSearchSinkConfiguration), any(TemplateStrategy.class), any()))
-                .thenReturn(indexManager);
-        doNothing().when(indexManager).setupIndex();
-        objectUnderTest.initialize();
-        
-        // Use ReflectivelySetField to set the dataStreamIndex
-        ReflectivelySetField.setField(OpenSearchSink.class, objectUnderTest, "dataStreamIndex", dataStreamIndex);
-        
-        // Verify the dataStreamIndex was set correctly
-        assertThat(objectUnderTest, notNullValue());
-    }
-    
-    @Test
-    void dataStreamIndex_ensureTimestamp_setsTimestampWhenMissing() throws Exception {
-        final OpenSearchSink objectUnderTest = createObjectUnderTest();
-        when(indexManagerFactory.getIndexManager(any(IndexType.class), eq(openSearchClient), any(RestHighLevelClient.class), eq(openSearchSinkConfiguration), any(TemplateStrategy.class), any()))
-                .thenReturn(indexManager);
-        doNothing().when(indexManager).setupIndex();
-        objectUnderTest.initialize();
-        
-        // Use ReflectivelySetField to set the dataStreamIndex
-        ReflectivelySetField.setField(OpenSearchSink.class, objectUnderTest, "dataStreamIndex", dataStreamIndex);
-        
-        // Verify the dataStreamIndex was set correctly
-        assertThat(objectUnderTest, notNullValue());
-    }
-
-    @ParameterizedTest
-    @MethodSource("externalVersionTypeProvider")
-    void initialize_sets_isExternalVersioning_true_on_BulkRetryStrategy_for_external_version_types(
-            final VersionType versionType) throws Exception {
-        when(indexConfiguration.getVersionType()).thenReturn(versionType);
-
-        final OpenSearchSink objectUnderTest = createObjectUnderTest();
-        when(indexManagerFactory.getIndexManager(any(IndexType.class), eq(openSearchClient), any(RestHighLevelClient.class), eq(openSearchSinkConfiguration), any(TemplateStrategy.class), any()))
-                .thenReturn(indexManager);
-        doNothing().when(indexManager).setupIndex();
-        objectUnderTest.initialize();
-
-        final BulkRetryStrategy bulkRetryStrategy = getField(objectUnderTest, "bulkRetryStrategy");
-        final boolean isExternalVersioning = getField(bulkRetryStrategy, "isExternalVersioning");
-        assertThat(isExternalVersioning, equalTo(true));
-    }
-
-    @ParameterizedTest
-    @MethodSource("nonExternalVersionTypeProvider")
-    void initialize_sets_isExternalVersioning_false_on_BulkRetryStrategy_for_non_external_version_types(
-            final VersionType versionType) throws Exception {
-        lenient().when(indexConfiguration.getVersionType()).thenReturn(versionType);
-
-        final OpenSearchSink objectUnderTest = createObjectUnderTest();
-        when(indexManagerFactory.getIndexManager(any(IndexType.class), eq(openSearchClient), any(RestHighLevelClient.class), eq(openSearchSinkConfiguration), any(TemplateStrategy.class), any()))
-                .thenReturn(indexManager);
-        doNothing().when(indexManager).setupIndex();
-        objectUnderTest.initialize();
-
-        final BulkRetryStrategy bulkRetryStrategy = getField(objectUnderTest, "bulkRetryStrategy");
-        final boolean isExternalVersioning = getField(bulkRetryStrategy, "isExternalVersioning");
-        assertThat(isExternalVersioning, equalTo(false));
-    }
-
-    private static Stream<Arguments> externalVersionTypeProvider() {
-        return Stream.of(
-                Arguments.of(VersionType.External),
-                Arguments.of(VersionType.ExternalGte)
-        );
-    }
-
-    private static Stream<Arguments> nonExternalVersionTypeProvider() {
-        return Stream.of(
-                Arguments.of(VersionType.Internal),
-                Arguments.of((VersionType) null)
-        );
+        verify(ingester).shutdown();
     }
 
     @SuppressWarnings("unchecked")
     private static <T> T getField(final Object target, final String fieldName) throws Exception {
-        Field field = target.getClass().getDeclaredField(fieldName);
+        java.lang.reflect.Field field = target.getClass().getDeclaredField(fieldName);
         field.setAccessible(true);
         return (T) field.get(target);
     }
